chore: Reorganized app provider logic
